Update Product.java
changed datatype on parameter in constructor

+++ src/model/Product.java
@@ -18,7 +18,7 @@ public class Product {
 		return price;
 	}
 	
-	public Product(String name, String description, int price) {
+	public Product(String name, String description, double price) {
 		this.name = name;
 		this.description = description;
 		this.price = price;
